gluster:Datatype change for bytesRead & bytesWritten from int to long.
This patch changes the datatype of bytesRead and bytesWritten from int
to long to avoid the data overflow that may occur.

Bug-url: https://bugzilla.redhat.com/show_bug.cgi?id=1015013

+++ backend/manager/modules/common/src/main/java/org/ovirt/engine/core/common/businessentities/gluster/GlusterClientInfo.java
@@ -14,9 +14,9 @@ public class GlusterClientInfo implements Serializable {
 
     private int clientPort;
 
-    private int bytesRead;
+    private long bytesRead;
 
-    private int bytesWritten;
+    private long bytesWritten;
 
     public String getHostname() {
         return hostname;
@@ -34,19 +34,19 @@ public class GlusterClientInfo implements Serializable {
         this.clientPort = clientPort;
     }
 
-    public int getBytesRead() {
+    public long getBytesRead() {
         return bytesRead;
     }
 
-    public void setBytesRead(int bytesRead) {
+    public void setBytesRead(long bytesRead) {
         this.bytesRead = bytesRead;
     }
 
-    public int getBytesWritten() {
+    public long getBytesWritten() {
         return bytesWritten;
     }
 
-    public void setBytesWritten(int bytesWritten) {
+    public void setBytesWritten(long bytesWritten) {
         this.bytesWritten = bytesWritten;
     }
 }
